feat(discover) Format durations in tooltips
When the current aggregate function is a duration function we should
format the tooltip value the same way we do in the table.
